+++ libexec/scoop-download.ps1
@@ -96,6 +96,7 @@ foreach ($app in $application) {
                     $title, $body = $_.Exception.Message -split '\|-'
                     if (!$body) { $body = $title }
                     Write-UserMessage -Message $body -Err
+                    debug $_.InvocationInfo
                     if ($title -ne 'Ignore' -and ($title -ne $body)) { New-IssuePrompt -Application $appName -Bucket $bucket -Title $title -Body $body }
 
                     continue
@@ -133,6 +134,7 @@ foreach ($app in $application) {
                         $title, $body = $_.Exception.Message -split '\|-'
                         if (!$body) { $body = $title }
                         Write-UserMessage -Message $body -Err
+                        debug $_.InvocationInfo
                         if ($title -ne 'Ignore' -and ($title -ne $body)) { New-IssuePrompt -Application $appName -Bucket $bucket -Title $title -Body $body }
 
                         continue