+++ libexec/scoop-install.ps1
@@ -182,6 +182,7 @@ foreach ($app in $apps) {
         $title, $body = $_.Exception.Message -split '\|-'
         if (!$body) { $body = $title }
         Write-UserMessage -Message $body -Err
+        debug $_.InvocationInfo
         if ($title -ne 'Ignore' -and ($title -ne $body)) { New-IssuePrompt -Application $cleanApp -Bucket $bucket -Title $title -Body $body }
 
         continue